Fix load to server

+++ app/src/main/java/com/goodlucky/finance/receipts/jsonObjects/Barcode.kt
@@ -75,7 +75,7 @@ data class Item(
     @SerializedName("name") val name : String,
     @SerializedName("ndsSum") val ndsSum : Double,
     @SerializedName("price") val price : Double,
-    @SerializedName("quantity") val quantity : Int,
+    @SerializedName("quantity") val quantity : Double,
     @SerializedName("sum") val sum : Double
 )
